major fixes to "allow-bare-refs"; implemented "assume-local" parse option to assume local scope for all variables if not otherwise declared; implemented "new-style" parse option to give both "allow-bare-refs" and "assume-local"; made references with "$" and "$." illegal when "allow-bare-refs" set
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@4015 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ include/qore/intern/qore_thread_intern.h
@@ -95,7 +95,7 @@ DLLLOCAL bool inMethod(const char *name, const QoreObject *o);
 DLLLOCAL void pushProgram(QoreProgram *pgm);
 DLLLOCAL void popProgram();
 DLLLOCAL RootQoreNamespace *getRootNS();
-DLLLOCAL int getParseOptions();
+DLLLOCAL int64 getParseOptions();
 DLLLOCAL void updateCVarStack(CVNode *ncvs);
 DLLLOCAL CVNode *getCVarStack();
 DLLLOCAL void updateVStack(VNode *nvs);
